feat: 근무 현황 페이지 추가

+++ src/main/java/com/itwillbs/factron/repository/work/WorkRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkRepository extends JpaRepository<WorkHistory, Integer> {
+public interface WorkRepository extends JpaRepository<WorkHistory, Long> {
 }
